+++ packages/web/src/modules/profiles/components/layout/snippets/ProfileSidebar.vue
@@ -107,8 +107,7 @@ function onMenuItemClick(item: IMenuItem) {
 
 <template>
   <nav v-if="isAuthenticated" id="sidebar" ref="sidebar" class="sidebar" :aria-label="ariaLabel">
-    <div
-      class="h-screen fix-h-screen sticky top-0 left-0 flex-col flex-wrap justify-start content-start items-start">
+    <div class="h- sticky top-0 left-0 flex-col flex-wrap justify-start content-start items-start">
       <div class="py-2">
         <a
           class="flex items-center no-underline font-extrabold uppercase tracking-wider h-12 px-3 cursor-pointer"